+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+
  * Bump MSRV to 1.64
+ * Update dependencies
 
 ## [0.4.2] - 2022-12-01