+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2023-02-07
+
+### Added
+
+ * Added torrent-csv as a new search provider.
+
 ## [0.4.2] - 2022-12-01
 
 ### Changed